+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.2
+
+- padding inside toast message
+- list of box shadow for toast message
+- toast message minimum/maximum height and width changed
+
 ## 1.0.1
 
 - changed dart sdk version